I defined a vector of all possible status codes from reply.h. From this, I defined a vector of all stock replies for every status code, as well as every buffer for all stock replies. That way, I could test all fields of all results with a single for loop in each test.
I moved some code from the namespaces in reply.cc to the namespaces in reply.h. This was done so the variables could be accessed in reply.h, and used in tests. I also think namespace variables are supposed to be in the header files by convention. I also declared some of the functions used in reply.cc in the header file to expose them to reply_test.cc. With these tests, between testing reply.cc and config_parser.cc, the code coverage stats are as follows

lines: 81.3% (230 out of 283)
branches: 54.1% (178 out of 329)

Change-Id: I853da69ab3c0a4edf2bc0a6dcecf8a3b1bea12e4