Made request_handler the base class, and created two derived classes: file_handler and echo_handler. They implement the pure virtual function "handle_request". In session.cc, when a request is sent to the request handler, it echoes if it does not parse as a valid request, and it handles the file request if it does parse. I removed reply::create_reply, as it now handled in handle_request and the usage seemed strange before (create a reply object and then call reply.create_reply..?)
Currently, the file_handler only handles .txt files. If the requested file exists, it returns successfully with the contents of the file. Otherwise, it returns a 404 error.

A "valid request" has been updated to refer to a request with a valid request line (3 tokens on the first line), where the first token must be GET (we can change this if more functionality is required), the second token is the path to a file (whose existence will be checked in file_handler) and the last token is HTTP/1.1. Headers may follow the request line.

To do:
Make "base directory" configurable
Handle other file types, like .html, .jpg, etc.
Support multiple servlets / handlers of the same type

Change-Id: I0a6df60b223eb5a8e2b58e2002dc42b1e63e09ed