add basic crud integration tests
These tests allow us to test that the program as a whole functions as expected. Currently, only create and retrieve are implemented, so this test only tests those functionalities. After the other operations are implemented, this test will be expanded.

This can be tested by using "make test" or any other equivalent testing command, as it has been added to CMakeLists.txt

Change-Id: I4be16947da354175ea524eda820d06a1afb54c5c