1/n added filesystem interface
created filesystem interface and implementation for persistent filesytem interface.
the interface is created for standardizing filesystem and would allow the creation of mock filesystem in the diffs to come (for testing purposes)

the diff is the first of the many diffs that'd create the filesystem for the repo. this task is split into multiple diffs for readability. future diffs would include modification of codebase for usage of this filesystem and creation of mock filesystem and unit tests.

tested using command: $ cd build && cmake ..; make; make test ARGS=-V
all tests passed -> no compilation error

Change-Id: I11f2a3f73411d098ae053f6ffbb1b3d9e580ccf8