2/n created filesystem instance for server
created persistence filesystem instance in server and passed them down in crud factory and handler. for this diff, the filesystem is initiated but not used for actual file system handling yet; the coming diff would do so.

patch 2 changes:
1. used obj ref instead of pointer for better memory mgmt
2. revised codebase and moved filesystem construction to server_main for better dependency injection

the diff is the second of the many diffs that'd create the filesystem for the repo. this task is split into multiple diffs for readability. future diffs would include modification of codebase for usage of this filesystem and creation of mock filesystem and unit tests.

tested using command: $ cd build && cmake ..; make; make test ARGS=-V
all tests passed -> no compilation error

Change-Id: I5c468fb9564404f3dbc3a27de4ebe2ff298e6649