What: Added functionality for parsing request body to get user move. I also updated the response body to include the updated fen string if the given move was successful. Tested that docker image builds and runs properly and that game is correctly instantiated. Will integrate this move checking functionality with frontend once html code for handling moving pieces on frontend is complete.
How: treated the user's request to move pieces on the board as a GET request and parsed through body of http request to look for '+' sign. It was agreed that the 4 characters after the '+' sign would be the starting and ending positions of the given piece. So, I extracted that information and tested the validity of the given move with the makeMove function from the chess API.

Update from patchset: updated the get_fen function so that it only returns the fen instead of the fen+move that it was currently returning. Also, added checks as per Max's suggestion for bounds on move parameters and length of move substring to prevent malicious requests from crashing server.

Change-Id: I736422bbf64950416ae78613484f98a293d0ea9f