Rebased Sahil's change -
Moved the AI move functionality to a separate function, handle_ai. On the path /FEN+ai, it calls this function, where it parses the FEN and tells the computer to makes a move. handle_ai doesn't perform much input validation, but only because it is only called after serve performs fen validation.

The move_pieces function now only takes in a FEN and makes the move posted after the "+" sign (/FEN+move). Instead of stoi, it now uses the - '0' method of converting a char to an int, which won't crash the server on letters, and will be filtered out when start/end row/col is being bound checked.

Testing: Retested functionality of /FEN, and /FEN+move. Tested the functionality of /FEN+ai, and checked bad move input, including illegal moves, letters, and moves not 4 characters long.

Change-Id: Id66d3ef0bdef2deae305885d783a5e255b3efa7f